backend/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.0; python_version >= '3.7'
dj-database-url==0.5.0
django-cors-headers==3.11.0
django-filter==21.1
django-heroku==0.3.1
django==4.0.4
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2-binary==2.9.3
psycopg2==2.9.3
pytz==2022.1
setuptools==62.1.0; python_version >= '3.7'
sqlparse==0.4.2; python_version >= '3.5'
whitenoise==6.0.0
